+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/customized/BootstrapOnlyProxyFactoryFactoryInitiator.java
@@ -10,6 +10,8 @@ import org.hibernate.service.spi.ServiceRegistryImplementor;
 /**
  * We need a different implementation of ProxyFactoryFactory during the build than at runtime,
  * so to allow metadata validation. This implementation is then swapped after the metadata has been recorded.
+ * 
+ * @see io.quarkus.hibernate.orm.runtime.customized.QuarkusRuntimeProxyFactoryFactory
  */
 public final class BootstrapOnlyProxyFactoryFactoryInitiator implements StandardServiceInitiator<ProxyFactoryFactory> {
